chore: Move LogHelper to modules directory
