+++ src/app/modules/LogHelper.ts
